Move team member related functions to a separate file and service
