+++ teams.go
@@ -8,7 +8,7 @@ import (
 	"fmt"
 )
 
-// TeamService handles communication with the teams related
+// TeamsService handles communication with the teams related
 // methods of the buildkite API.
 //
 // buildkite API docs: https://buildkite.com/docs/api
@@ -43,25 +43,6 @@ type CreateTeam struct {
 	IsDefaultTeam             bool   `json:"is_default_team,omitempty"`
 	DefaultMemberRole         string `json:"default_member_role,omitempty"`
 	MembersCanCreatePipelines bool   `json:"members_can_create_pipelines,omitempty"`
-}
-
-// TeamMember represents a member of a team.
-type TeamMember struct {
-	ID        string    `json:"user_id,omitempty"`
-	UserName  string    `json:"user_name,omitempty"`
-	Role      string    `json:"role,omitempty"`
-	CreatedAt Timestamp `json:"created_at,omitempty"`
-}
-
-// CreateTeamMember represents a request to create a team member.
-type CreateTeamMember struct {
-	UserID string `json:"user_id,omitempty"`
-	Role   string `json:"role,omitempty"`
-}
-
-// TeamMembersListOptions specifies the optional parameters to the ListTeamMembers method.
-type TeamMembersListOptions struct {
-	ListOptions
 }
 
 // Get the teams for an org.
@@ -145,99 +126,6 @@ func (ts *TeamsService) UpdateTeam(ctx context.Context, org string, id string, t
 // DeleteTeam deletes a team.
 func (ts *TeamsService) DeleteTeam(ctx context.Context, org string, id string) (*Response, error) {
 	u := fmt.Sprintf("v2/organizations/%s/teams/%s", org, id)
-
-	req, err := ts.client.NewRequest(ctx, "DELETE", u, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	resp, err := ts.client.Do(req, nil)
-	if err != nil {
-		return resp, err
-	}
-
-	return resp, err
-}
-
-// ListTeamMembers lists the members of a team.
-func (ts *TeamsService) ListTeamMembers(ctx context.Context, org string, id string, opt *TeamMembersListOptions) ([]TeamMember, *Response, error) {
-	u := fmt.Sprintf("v2/organizations/%s/teams/%s/members", org, id)
-	u, err := addOptions(u, opt)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	req, err := ts.client.NewRequest(ctx, "GET", u, nil)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	var teamMembers []TeamMember
-	resp, err := ts.client.Do(req, &teamMembers)
-	if err != nil {
-		return nil, resp, err
-	}
-
-	return teamMembers, resp, err
-}
-
-// GetTeamMember gets a team member.
-func (ts *TeamsService) GetTeamMember(ctx context.Context, org string, teamID string, userID string) (TeamMember, error) {
-	u := fmt.Sprintf("v2/organizations/%s/teams/%s/members/%s", org, teamID, userID)
-
-	req, err := ts.client.NewRequest(ctx, "GET", u, nil)
-	if err != nil {
-		return TeamMember{}, err
-	}
-
-	var teamMember TeamMember
-	_, err = ts.client.Do(req, &teamMember)
-	if err != nil {
-		return teamMember, err
-	}
-
-	return teamMember, err
-}
-
-// CreateTeamMember creates a team member.
-func (ts *TeamsService) CreateTeamMember(ctx context.Context, org string, teamID string, t CreateTeamMember) (TeamMember, *Response, error) {
-	u := fmt.Sprintf("v2/organizations/%s/teams/%s/members", org, teamID)
-
-	req, err := ts.client.NewRequest(ctx, "POST", u, t)
-	if err != nil {
-		return TeamMember{}, nil, err
-	}
-
-	var teamMember TeamMember
-	resp, err := ts.client.Do(req, &teamMember)
-	if err != nil {
-		return teamMember, resp, err
-	}
-
-	return teamMember, resp, err
-}
-
-// UpdateTeamMember updates a team member.
-func (ts *TeamsService) UpdateTeamMember(ctx context.Context, org string, teamID string, userID string, role string) (TeamMember, *Response, error) {
-	u := fmt.Sprintf("v2/organizations/%s/teams/%s/members/%s", org, teamID, userID)
-
-	req, err := ts.client.NewRequest(ctx, "PATCH", u, map[string]string{"role": role})
-	if err != nil {
-		return TeamMember{}, nil, err
-	}
-
-	var teamMember TeamMember
-	resp, err := ts.client.Do(req, &teamMember)
-	if err != nil {
-		return teamMember, resp, err
-	}
-
-	return teamMember, resp, err
-}
-
-// DeleteTeamMember deletes a team member.
-func (ts *TeamsService) DeleteTeamMember(ctx context.Context, org string, teamID string, userID string) (*Response, error) {
-	u := fmt.Sprintf("v2/organizations/%s/teams/%s/members/%s", org, teamID, userID)
 
 	req, err := ts.client.NewRequest(ctx, "DELETE", u, nil)
 	if err != nil {
